Add celery and rename commands to tasks
