+++ deps/base.pip.txt
@@ -1,4 +1,6 @@
+celery >= 5.0.5, < 5.1
 django-bootstrap4
+django-celery-beat >= 2.2.0, < 2.3
 django-environ
 django-redis
 django-tinymce
